statistical analysis on bed files
These are some code for some statistical analysis and machien learnign mdoelign on bed files. The code dfinitely has to be revised as there are hard coded paths and queries to the data files that do not exist in the repo yet. Ali and Ankita will revise the codes to make sure the code will be revised properly and repushed to the repo.